neural_backend/requirements.txt
# Core ML and Data Processing
tensorflow==2.13.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Backend Framework
flask==2.3.2
flask-cors==4.0.0
flask-limiter==3.5.0
flask-jwt-extended==4.5.2

# Speech Processing
librosa==0.10.1
soundfile==0.12.1
speechrecognition==3.10.0
pyttsx3==2.90
gtts==2.3.2
pydub==0.25.1

# API and Documentation
flask-restx==1.1.0
flask-swagger-ui==4.11.1

# Security and Authentication
bcrypt==4.0.1
python-dotenv==1.0.0
cryptography==41.0.3

# Logging and Monitoring
python-json-logger==2.0.7
prometheus-flask-exporter==0.22.4

# Data Validation
marshmallow==3.20.1
pydantic==2.3.0

# Database (optional)
sqlalchemy==2.0.19
flask-sqlalchemy==3.0.5

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Development
black==23.7.0
flake8==6.0.0
mypy==1.5.1

# Containerization
gunicorn==21.2.0
gevent==23.7.0